# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My CMS API",
        "version": "0.1"
    },
    "paths": {
...
        "/postseo": {
            "get": {
                "tags": [
                    "PostSeo"
                ],
                "summary": "Get list of posts seo",
                "description": "Returns list of posts seo",
                "operationId": "getPostSeosList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "PostSeo"
                ],
                "summary": "Create a post seo",
                "description": "Create a post seo",
                "operationId": "createAPostSeo",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/postseo/:id": {
            "get": {
                "tags": [
                    "PostSeo"
                ],
                "summary": "Get a post seo",
                "operationId": "getAPostSeo",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "PostSeo"
                ],
                "summary": "Update a post seo",
                "operationId": "updateAPostSeo",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "PostSeo"
                ],
                "summary": "Delete a post seo",
                "operationId": "deleteAPostSeo",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/setting": {
            "get": {
                "tags": [
                    "Setting"
                ],
                "summary": "Get list of settings",
                "description": "Returns list of settings",
                "operationId": "getSettingsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Setting"
                ],
                "summary": "Create a setting",
                "description": "Create a setting",
                "operationId": "createASetting",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/setting/:id": {
            "get": {
                "tags": [
                    "Setting"
                ],
                "summary": "Get a setting",
                "operationId": "getASetting",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Setting"
                ],
                "summary": "Update a setting",
                "operationId": "updateASetting",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Setting"
                ],
                "summary": "Delete a setting",
                "operationId": "deleteASetting",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    },
...
}